Merge the functionality of assert_abort() and assert_static_abort() into assert_abort().
Print the unexpanded message to klog first. Then try to print the
message and the stack trace to standard output using printf. Use the
same unabbreviated message wording for both klog and stdout.  If a
nested or a parallel assert is detected, send the message only to klog
and then abort.
      
Re-introduce malloc_assert() to up the malloc_futex. This appears to be
inevitable if we want to print to stdout. To prevent undesired macro
expansion, call assert_abort() from malloc_assert() directly.