libposix: use objcopy more effectively
This removes the helper libposix tools that were used to redefine
conflicting symbols in libc.

objcopy can work on whole libraries thus there is no need to
unpack and repack them again.
Also, it accepts file with list of redefinitions, thus we do not
need to use xargs at all.

This solution is much simpler and also it speeds-up the compilation
a bit.