Read toolchain patches from the correct directory.
When running toolchain.sh with current directory other than tools
subdirectory of the HelenOS sources, the script was unable to
correctly read toolchain patches, as it was trying to read them
from BASEDIR (i.e. the current working directory when running
the script).

This changeset explicitly uses path to the patches relative to
the location of the directory where the toolchain script resides
(the newly introduced SRCDIR variable).