Fix L_CREATE and L_UNLINK across mountpoints
Server-side mountpoints broke linking of the first name and unlinking of
the last name of a file across mountpoints. In order to do the expected
thing, both operations need to be split into a mere lookup phase to
locate the parent of the added/removed name, and a link/unlink phase
that adds/removes the name to/from the parent node.