Set logger's VFS root
As an init task, logger starts with no VFS root, which effectively
disables logging to /log/*. In this commit we add a way to set logger's
root after the root is mounted in init to fix this. Note that some log
entries might be still missing as file-based logging kicks in quite
late.